add book removal functionality
Using the data-attribute, each book card was assigned the
book's id. Rather than assigning an event listener to each
card, a single event listener was attached to the library
grid, which listens for a click on the delete book button.

A new function removeBookFromLibrary() was created to locate
the book to be removed, and rerender the library display.